Refactor Course Chat Application: Integrate LangChain for Chat Functionality, Update Tools, and Revise Prompts
- Replaced the existing agent initialization with LangChain's chat model integration in app.py.
- Simplified chat function to utilize LangChain's message handling.
- Updated main.py to define a new graph structure using LangGraph for improved tool management.
- Revised prompts to reflect a more engaging and focused assistant persona.
- Enhanced animal fact retrieval methods for better error handling and response formatting.
- Streamlined music recommendation logic and integrated additional details fetching from a SQL database.
- Updated README to reflect changes in architecture and usage instructions.
