+++ 05_src/course_agents/last_path.py
